+++ pkg/release/changelog/changelog.go
@@ -172,6 +172,12 @@ var KnownComponents = map[string]ParseParams{
 		Start:     commonStartPattern,
 		End:       commonEndPattern,
 	},
+	"kube-vip-cloud-provider": {
+		Tag:       "https://github.com/giantswarm/kube-vip-cloud-provider-app/releases/tag/v{{.Version}}",
+		Changelog: "https://raw.githubusercontent.com/giantswarm/kube-vip-cloud-provider-app/v{{.Version}}/CHANGELOG.md",
+		Start:     commonStartPattern,
+		End:       commonEndPattern,
+	},
 	"vsphere-csi-driver": {
 		Tag:       "https://github.com/giantswarm/vsphere-csi-driver-app/releases/tag/v{{.Version}}",
 		Changelog: "https://raw.githubusercontent.com/giantswarm/vsphere-csi-driver-app/v{{.Version}}/CHANGELOG.md",